Add test for nesting subrules.
This is a major problem, and is the only thing (AFAICT) that's keeping
me from putting it in the ecosystem. I _think_ the issue involves the
creation of new Cursors per subrule, and changes being made to that
Cursor's Parsefail role not propagating back upwards. May have to create
my own Cursor to solve that.